rename integration back to OpenAI (not v2)
