--- Packs/OpenAI/Integrations/OpenAIv2/OpenAIv2_description.md
@@ -1,27 +1,0 @@
-## OpenAI
-- Create an API key on https://platform.openai.com/: 
-  - Open user account settings (at upper right)
-  - Click "View API Keys"
-  - Click "+ Create new secret key"
-- Use default value https://api.openai.com/ as Server URL
-- Enter OpenAI API key in API Key
-- All other settings can be left as the default
-
-## Azure OpenAI
-- [Create an API key in Azure OpenAI Resources](https://learn.microsoft.com/en-us/azure/cognitive-services/openai/quickstart?tabs=command-line&pivots=rest-api)
-- Use value https://<your-resource-name>.openai.azure.com/ as Server URL
-- Enter Azure OpenAI API key in API Key
-- Check the box "Use Azure OpenAI?"
-- Optionally, enter a different API Version
-- Enter a different Model to Use for Test if the default model, "text-davinci-003", is not deployed
-
-## XSOAR Use Cases
-[Playbook of the Week: Using ChatGPT in Cortex XSOAR](https://www.paloaltonetworks.com/blog/security-operations/using-chatgpt-in-cortex-xsoar/)
-
----
-The following Description is generated by ChatGPT Integration:
-OpenAI ChatGPT is a revolutionary conversational AI system that is designed to interact with humans just like real people. Developed by one of the most innovative and respected names in the field of artificial intelligence, this cutting-edge system is capable of understanding human language, responding to complex queries, and even mimicking human personality traits. Using advanced natural language processing techniques, OpenAI ChatGPT can simulate human conversation to an unprecedented degree of accuracy and effectiveness, making it an ideal tool for a wide range of applications, including customer service, virtual assistance, and more. Whether you are looking for a smarter way to engage with your customers or simply want to explore the cutting edge of AI technology, OpenAI ChatGPT is the perfect solution for all your conversational needs.
-
----
-OpenAI - [View Integration Documentation](https://xsoar.pan.dev/docs/reference/integrations/open-ai)
-OpenAi ChatGPT v3 - [View Integration Documentation](https://xsoar.pan.dev/docs/reference/integrations/open-ai-chat-gpt-v3)
